fix layering violations in component updater and add missing dep fix https://github.com/brave/brave-browser/issues/5744
