+++ browser/brave_browser_process_impl.h
@@ -25,6 +25,9 @@ class BraveWidevineBundleManager;
 #endif
 
 namespace brave_component_updater {
+#if BUILDFLAG(ENABLE_EXTENSIONS)
+class ExtensionWhitelistService;
+#endif
 class LocalDataFilesService;
 }
 
@@ -33,7 +36,6 @@ class AdBlockService;
 class AdBlockCustomFiltersService;
 class AdBlockRegionalServiceManager;
 class AutoplayWhitelistService;
-class ExtensionWhitelistService;
 class HTTPSEverywhereService;
 class ReferrerWhitelistService;
 class TrackingProtectionService;
@@ -55,7 +57,6 @@ class BraveBrowserProcessImpl : public BrowserProcessImpl {
   ~BraveBrowserProcessImpl() override;
 
   // BrowserProcess implementation.
-  component_updater::ComponentUpdateService* component_updater() override;
   void ResourceDispatcherHostCreated() override;
 
   ProfileManager* profile_manager() override;
@@ -66,7 +67,8 @@ class BraveBrowserProcessImpl : public BrowserProcessImpl {
   ad_block_regional_service_manager();
   brave_shields::AutoplayWhitelistService* autoplay_whitelist_service();
 #if BUILDFLAG(ENABLE_EXTENSIONS)
-  brave_shields::ExtensionWhitelistService* extension_whitelist_service();
+  brave_component_updater::ExtensionWhitelistService*
+  extension_whitelist_service();
 #endif
   brave_shields::ReferrerWhitelistService* referrer_whitelist_service();
   greaselion::GreaselionDownloadService* greaselion_download_service();
@@ -98,8 +100,10 @@ class BraveBrowserProcessImpl : public BrowserProcessImpl {
       ad_block_regional_service_manager_;
   std::unique_ptr<brave_shields::AutoplayWhitelistService>
       autoplay_whitelist_service_;
-  std::unique_ptr<brave_shields::ExtensionWhitelistService>
+#if BUILDFLAG(ENABLE_EXTENSIONS)
+  std::unique_ptr<brave_component_updater::ExtensionWhitelistService>
       extension_whitelist_service_;
+#endif
   std::unique_ptr<brave_shields::ReferrerWhitelistService>
       referrer_whitelist_service_;
   std::unique_ptr<greaselion::GreaselionDownloadService>
