+++ browser/component_updater/brave_component_updater_configurator.h
@@ -1,11 +1,17 @@
-/* This Source Code Form is subject to the terms of the Mozilla Public
+/* Copyright (c) 2019 The Brave Authors. All rights reserved.
+ * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-#ifndef BRAVE_BROWSER_COMPONENT_UPDATER_CHROME_COMPONENT_UPDATER_CONFIGURATOR_H_
-#define BRAVE_BROWSER_COMPONENT_UPDATER_CHROME_COMPONENT_UPDATER_CONFIGURATOR_H_
+#ifndef BRAVE_BROWSER_COMPONENT_UPDATER_BRAVE_COMPONENT_UPDATER_CONFIGURATOR_H_
+#define BRAVE_BROWSER_COMPONENT_UPDATER_BRAVE_COMPONENT_UPDATER_CONFIGURATOR_H_
+
+#include <memory>
+#include <string>
+#include <vector>
 
 #include "base/memory/ref_counted.h"
+#include "components/component_updater/configurator_impl.h"
 #include "components/update_client/configurator.h"
 
 class PrefRegistrySimple;
@@ -21,18 +27,55 @@ class URLRequestContextGetter;
 
 namespace component_updater {
 
-// Registers preferences associated with the component updater configurator
-// for Chrome. The preferences must be registered with the local pref store
-// before they can be queried by the configurator instance.
-// This function is called before MakeChromeComponentUpdaterConfigurator.
-void RegisterPrefsForBraveComponentUpdaterConfigurator(
-    PrefRegistrySimple* registry);
+class BraveConfigurator : public update_client::Configurator {
+ public:
+  BraveConfigurator(const base::CommandLine* cmdline,
+                    PrefService* pref_service);
 
-scoped_refptr<update_client::Configurator>
-MakeBraveComponentUpdaterConfigurator(
-    const base::CommandLine* cmdline,
-    PrefService* pref_service);
+  // update_client::Configurator overrides.
+  int InitialDelay() const override;
+  int NextCheckDelay() const override;
+  int OnDemandDelay() const override;
+  int UpdateDelay() const override;
+  std::vector<GURL> UpdateUrl() const override;
+  std::vector<GURL> PingUrl() const override;
+  std::string GetProdId() const override;
+  base::Version GetBrowserVersion() const override;
+  std::string GetChannel() const override;
+  std::string GetBrand() const override;
+  std::string GetLang() const override;
+  std::string GetOSLongName() const override;
+  base::flat_map<std::string, std::string> ExtraRequestParams() const override;
+  std::string GetDownloadPreference() const override;
+  scoped_refptr<update_client::NetworkFetcherFactory> GetNetworkFetcherFactory()
+      override;
+  scoped_refptr<update_client::UnzipperFactory> GetUnzipperFactory() override;
+  scoped_refptr<update_client::PatcherFactory> GetPatcherFactory() override;
+  bool EnabledDeltas() const override;
+  bool EnabledComponentUpdates() const override;
+  bool EnabledBackgroundDownloader() const override;
+  bool EnabledCupSigning() const override;
+  PrefService* GetPrefService() const override;
+  update_client::ActivityDataService* GetActivityDataService() const override;
+  bool IsPerUserInstall() const override;
+  std::vector<uint8_t> GetRunActionKeyHash() const override;
+  std::string GetAppGuid() const override;
+  std::unique_ptr<update_client::ProtocolHandlerFactory>
+  GetProtocolHandlerFactory() const override;
+  update_client::RecoveryCRXElevator GetRecoveryCRXElevator() const override;
+
+ private:
+  friend class base::RefCountedThreadSafe<BraveConfigurator>;
+
+  ConfiguratorImpl configurator_impl_;
+  PrefService* pref_service_;  // This member is not owned by this class.
+  scoped_refptr<update_client::NetworkFetcherFactory> network_fetcher_factory_;
+  scoped_refptr<update_client::UnzipperFactory> unzip_factory_;
+  scoped_refptr<update_client::PatcherFactory> patch_factory_;
+
+  ~BraveConfigurator() override;
+};
 
 }  // namespace component_updater
 
-#endif  // BRAVE_BROWSER_COMPONENT_UPDATER_CHROME_COMPONENT_UPDATER_CONFIGURATOR_H_
+#endif  // BRAVE_BROWSER_COMPONENT_UPDATER_BRAVE_COMPONENT_UPDATER_CONFIGURATOR_H_
