+++ browser/extensions/brave_extension_provider.cc
@@ -13,7 +13,7 @@
 #include "brave/browser/brave_browser_process_impl.h"
 #include "brave/common/extensions/extension_constants.h"
 #include "brave/components/brave_component_updater/browser/local_data_files_service.h"
-#include "brave/components/brave_shields/browser/extension_whitelist_service.h"
+#include "brave/components/brave_component_updater/browser/extension_whitelist_service.h"
 #include "brave/grit/brave_generated_resources.h"
 #include "ui/base/l10n/l10n_util.h"
 
@@ -40,54 +40,6 @@ bool IsBlacklisted(const extensions::Extension* extension) {
 }  // namespace
 
 namespace extensions {
-
-bool BraveExtensionProvider::IsVetted(const Extension* extension) {
-  // This is a hardcoded list of vetted extensions, mostly
-  // the built-in ones that ship with Brave or are used for
-  // unit tests.
-  // Don't add new extensions to this list. Add them to
-  // the files managed by the extension whitelist service.
-  return BraveExtensionProvider::IsVetted(extension->id());
-}
-
-bool BraveExtensionProvider::IsVetted(const std::string id) {
-  static std::vector<std::string> vetted_extensions({
-      brave_extension_id,
-      brave_rewards_extension_id,
-      brave_sync_extension_id,
-      brave_webtorrent_extension_id,
-      crl_set_extension_id,
-      ethereum_remote_client_extension_id,
-      hangouts_extension_id,
-      widevine_extension_id,
-      brave_component_updater::kLocalDataFilesComponentId,
-      // Web Store
-      "ahfgeienlihckogmohjhadlkjgocpleb",
-      // Brave Automation Extension
-      "aapnijgdinlhnhlmodcfapnahmbfebeb",
-      // Test ID: Brave Default Ad Block Updater
-      "naccapggpomhlhoifnlebfoocegenbol",
-      // Test ID: Brave Regional Ad Block Updater
-      // (9852EFC4-99E4-4F2D-A915-9C3196C7A1DE)
-      "dlpmaigjliompnelofkljgcmlenklieh",
-      // Test ID: Brave Tracking Protection Updater
-      "eclbkhjphkhalklhipiicaldjbnhdfkc",
-      // Test ID: PDFJS
-      "kpbdcmcgkedhpbcpfndimofjnefgjidd",
-      // Test ID: Brave HTTPS Everywhere Updater
-      "bhlmpjhncoojbkemjkeppfahkglffilp",
-      // Test ID: Brave Tor Client Updater
-      "ngicbhhaldfdgmjhilmnleppfpmkgbbk",
-      // Chromium PDF Viewer.
-      "mhjfbmdgcfjbbpaeojofohoefgiehjai",
-  });
-  if (std::find(vetted_extensions.begin(), vetted_extensions.end(), id) !=
-      vetted_extensions.end())
-    return true;
-
-  return g_brave_browser_process->extension_whitelist_service()->IsWhitelisted(
-      id);
-}
 
 BraveExtensionProvider::BraveExtensionProvider() {}
