+++ common/extensions/whitelist.h
@@ -1,0 +1,14 @@
+/* Copyright (c) 2019 The Brave Authors. All rights reserved.
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this file,
+ * You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+#ifndef BRAVE_COMMON_EXTENSIONS_WHITELIST_H_
+#define BRAVE_COMMON_EXTENSIONS_WHITELIST_H_
+
+#include <string>
+#include <vector>
+
+extern const std::vector<std::string> kVettedExtensions;
+
+#endif  // BRAVE_COMMON_EXTENSIONS_WHITELIST_H_
