+++ components/brave_component_updater/browser/extension_whitelist_service.h
@@ -3,8 +3,8 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-#ifndef BRAVE_COMPONENTS_BRAVE_SHIELDS_BROWSER_EXTENSION_WHITELIST_SERVICE_H_
-#define BRAVE_COMPONENTS_BRAVE_SHIELDS_BROWSER_EXTENSION_WHITELIST_SERVICE_H_
+#ifndef BRAVE_COMPONENTS_BRAVE_COMPONENT_UPDATER_BROWSER_EXTENSION_WHITELIST_SERVICE_H_
+#define BRAVE_COMPONENTS_BRAVE_COMPONENT_UPDATER_BROWSER_EXTENSION_WHITELIST_SERVICE_H_
 
 #include <map>
 #include <memory>
@@ -22,11 +22,11 @@ class ExtensionWhitelistParser;
 class BraveExtensionProviderTest;
 class BravePDFDownloadTest;
 
-using brave_component_updater::LocalDataFilesObserver;
-using brave_component_updater::LocalDataFilesService;
+namespace extensions {
+class Extension;
+}
 
-// TODO(bridiver) - move out of brave shields
-namespace brave_shields {
+namespace brave_component_updater {
 
 // The brave shields service in charge of extension whitelist
 class ExtensionWhitelistService : public LocalDataFilesObserver {
@@ -35,11 +35,14 @@ class ExtensionWhitelistService : public LocalDataFilesObserver {
       brave_component_updater::LoadDATFileDataResult<ExtensionWhitelistParser>;
 
   explicit ExtensionWhitelistService(
-      LocalDataFilesService* local_data_files_service);
+      LocalDataFilesService* local_data_files_service,
+      const std::vector<std::string>& whitelist);
   ~ExtensionWhitelistService() override;
 
   bool IsWhitelisted(const std::string& extension_id) const;
   bool IsBlacklisted(const std::string& extension_id) const;
+  bool IsVetted(const std::string& extension_id) const;
+  bool IsVetted(const extensions::Extension* extension) const;
 
   // implementation of LocalDataFilesObserver
   void OnComponentReady(const std::string& component_id,
@@ -52,9 +55,10 @@ class ExtensionWhitelistService : public LocalDataFilesObserver {
 
   void OnGetDATFileData(GetDATFileDataResult result);
 
+  SEQUENCE_CHECKER(sequence_checker_);
   std::unique_ptr<ExtensionWhitelistParser> extension_whitelist_client_;
   brave_component_updater::DATFileDataBuffer buffer_;
-  SEQUENCE_CHECKER(sequence_checker_);
+  std::vector<std::string> whitelist_;
   base::WeakPtrFactory<ExtensionWhitelistService> weak_factory_;
 
   DISALLOW_COPY_AND_ASSIGN(ExtensionWhitelistService);
@@ -62,8 +66,9 @@ class ExtensionWhitelistService : public LocalDataFilesObserver {
 
 // Creates the ExtensionWhitelistService
 std::unique_ptr<ExtensionWhitelistService> ExtensionWhitelistServiceFactory(
-    LocalDataFilesService* local_data_files_service);
+    LocalDataFilesService* local_data_files_service,
+    const std::vector<std::string>& whitelist);
 
-}  // namespace brave_shields
+}  // namespace brave_component_updater
 
-#endif  // BRAVE_COMPONENTS_BRAVE_SHIELDS_BROWSER_EXTENSION_WHITELIST_SERVICE_H_
+#endif  // BRAVE_COMPONENTS_BRAVE_COMPONENT_UPDATER_BROWSER_EXTENSION_WHITELIST_SERVICE_H_
