--- patches/components-component_updater-component_updater_service_internal.h.patch
@@ -1,20 +1,0 @@
-diff --git a/components/component_updater/component_updater_service_internal.h b/components/component_updater/component_updater_service_internal.h
-index 341fbf78a96bb1b0b8e10bf9579dea9b930bec13..5d1d80b412cc8669c0acfd3b707651af27b62809 100644
---- a/components/component_updater/component_updater_service_internal.h
-+++ b/components/component_updater/component_updater_service_internal.h
-@@ -26,6 +26,7 @@ enum class Error;
- 
- namespace component_updater {
- 
-+class BraveCrxUpdateService;
- class OnDemandUpdater;
- 
- using CrxInstaller = update_client::CrxInstaller;
-@@ -66,6 +67,7 @@ class CrxUpdateService : public ComponentUpdateService,
-                       Callback callback) override;
- 
-  private:
-+  friend class BraveCrxUpdateService;
-   void Start();
-   void Stop();
-
